Minor warning fixes

+++ rpcs3/Emu/RSX/RSXThread.cpp
@@ -2645,7 +2645,7 @@ namespace rsx
 		case frame_limit_type::_50: limit = 50.; break;
 		case frame_limit_type::_60: limit = 60.; break;
 		case frame_limit_type::_30: limit = 30.; break;
-		case frame_limit_type::_auto: limit = g_cfg.video.vblank_rate; break; // TODO
+		case frame_limit_type::_auto: limit = static_cast<double>(g_cfg.video.vblank_rate); break;
 		default:
 			break;
 		}
